requirements.txt
aiogram>=3.0.0
python-dotenv>=1.0.0
python-telegram-bot>=20.0
aiomysql>=0.2.0
cryptography>=42.0.0
aiojobs>=1.2.0
pymysql>=1.1.0